requirements.txt
fastapi
uvicorn[standard]
faster-whisper
python-multipart
python-dotenv
groq
openai
# Apple Silicon only (optional): pip install mlx-whisper
